2.0第一次

+++ packages/org.opencomb.oauth/auth/AuthoritionObtaining.php
@@ -1,6 +1,8 @@
 <?php
 namespace org\opencomb\oauth\auth ;
 
+use org\opencomb\coresystem\mvc\controller\UserSpace;
+
 use org\opencomb\coresystem\user\UserPanel;
 use org\jecat\framework\db\ExecuteException;
 use org\opencomb\coresystem\auth\Id;
@@ -11,39 +13,13 @@ use org\opencomb\oauth\adapter\AdapterManager;
 use org\jecat\framework\session\Session;
 use org\jecat\framework\message\Message;
 
-class AuthoritionObtaining extends UserPanel
+class AuthoritionObtaining extends UserSpace
 {
-	public function createBeanConfig()
-	{
-		return array(
+	protected $arrConfig = array(
 			'view' => array(
 					'template'=>'oauth:auth/AuthoritionObtaining.html' ,
-					'model' => 'user' ,
-					'class' => 'form' ,
-					'hideForm' => true ,
 			) ,
-			'model:user' => array(
-					'orm'=>array(
-							'table' => 'coresystem:user' ,
-							'hasOne:info' => array(
-								'table' => 'coresystem:userinfo'
-							) ,
-							'hasOne:token' => array(
-								'table' => 'oauth:user' ,
-								'keys' => array('service','suid') ,
-								'fromkeys' => 'uid' ,
-								'tokeys' => 'uid' ,
-							) ,
-					)
-			) ,
-			'model:token' => array(
-					'orm'=>array(
-							'table' => 'oauth:user' ,
-							'keys' => array('service','suid') ,
-					)
-			) ,
-		) ;
-	}
+	) ;	
 	
 	public function process()
 	{
@@ -52,12 +28,18 @@ class AuthoritionObtaining extends UserPanel
 		{
 			$this->params['act'] = 'form' ;
 		}
+		
+		$this->model('coresystem:user')
+		     ->hasOne('coresystem:userinfo','uid','uid')
+		     ->hasOne('oauth:user','uid','uid','token');
+		
+		$this->model('oauth:user');
+		
 		$this->doActions() ;
 	}
 	
-	protected function actionForm()
+	protected function form()
 	{
-	    
 		// 检查参数 ----------
 		if( empty($this->params['service']) or (empty($this->params['oauth_token']) and (empty($this->params['code']))) )
 		{
@@ -115,34 +97,38 @@ class AuthoritionObtaining extends UserPanel
 		}
 		
 		// 检查 token 是否存在
-		$this->user->load(
-				array($this->params['service'],$this->arrAccessToken['id'])
-				, array('token.service','token.suid')
-		) ;
+		$this->model('coresystem:user')->where("token.service = '{$this->params['service']}'");
+		$this->model('coresystem:user')->where("token.suid = '{$this->arrAccessToken['id']}'");
+		$this->model('coresystem:user')->load() ;
 		
 		if( $this->params['operation']=='login' )
 		{
 			// 直接登录
-			if( !$this->user->isEmpty() )
+			if( $this->model('coresystem:user')->rowNum() > 0 )
 			{
-				if( !IdManager::singleton()->id($this->user->uid) )
+				if( !IdManager::singleton()->id($this->model('coresystem:user')->uid) )
 				{
 					IdManager::singleton()->setCurrentId(
-						new Id($this->user)
+						new Id($this->model('coresystem:user'))
 					) ;
 				}
 				
 				
-				if($this->user->child('token')->token && $this->user->child('token')->token_secret)
+				if($this->model('coresystem:user')->data('token.token') && $this->model('coresystem:user')->data('token.token_secret'))
 				{
-				    $this->createMessage( Message::success, '已经以 %s 的身份登录', $this->user->username ) ;
+				    $this->createMessage( Message::success, '已经以 %s 的身份登录', $this->model('coresystem:user')->data('username') ) ;
 				}else{
 				    
-				    $this->user->setData("token.token",$this->arrAccessToken['oauth_token']);
-				    $this->user->setData("token.token_secret",$this->arrAccessToken['oauth_token_secret']);
+				    $this->model('coresystem:user')->setData("token.token",$this->arrAccessToken['oauth_token']);
+				    $this->model('coresystem:user')->setData("token.token_secret",$this->arrAccessToken['oauth_token_secret']);
 				    
-				    $this->user->save();
-				    $this->createMessage( Message::success, '已经以 %s 的身份绑定并登录', $this->user->username ) ;
+				    $this->model('coresystem:user')->insert(
+				            array(
+				                    "token.token" => $this->arrAccessToken['oauth_token'],
+				                    "token.token_secret" => $this->arrAccessToken['oauth_token_secret']
+		                    )
+		            );
+				    $this->createMessage( Message::success, '已经以 %s 的身份绑定并登录', $this->model('coresystem:user')->username ) ;
 				}
 				
 				
@@ -165,12 +151,14 @@ class AuthoritionObtaining extends UserPanel
 		{
 			$this->view->variables()->set('sBindUser',$aId->username()) ;
 		}
+		
+		$this->view()->setModel($this->model('coresystem:user'));
 	}
 	
 	/**
 	 * 绑定已有的帐号
 	 */
-	protected function actionBindExists()
+	protected function BindExists()
 	{
 		$arrAccessToken = Session::singleton()->variable($this->params['service'].'.AccessToken') ;
 		
@@ -182,27 +170,27 @@ class AuthoritionObtaining extends UserPanel
 		}
 		
 		
-		if( !$this->user->load( $this->params['user'],'username') )
+		if( !$this->model('coresystem:user')->load( $this->params['user'],'username') )
 		{
 			$this->createMessage(Message::error,"用户名不存在，无法完成绑定") ;
 			return ;
 		}
 		
-		if( $this->user['password']!=Authenticate::encryptPassword($this->user,$this->params['user'],$this->params['password']) )
+		if( $this->model('coresystem:user')->data('password')!=Authenticate::encryptPassword($this->model('coresystem:user'),$this->params['user'],$this->params['password']) )
 		{
 			$this->createMessage(Message::error,"密码错误，无法完成绑定") ;
 			return ;
 		}
 		
 		// 解除以前的绑定
-		$aBind = $this->user->child('token')->prototype()->createModel() ;
-		if( $aBind->load(
-				array($this->params['service'],$arrAccessToken['id'])
-				/* , array('service','suid') */
-		) )
+		$aBind = $this->model('oauth:user') ;
+		$aBind->where("service = '{$this->params['service']}'");
+		$aBind->where("suid = '{$arrAccessToken['id']}'");
+		
+		if( $aBind->load()->rowNum() > 0 )
 		{
 			try{
-				$aBind->delete() ;
+				$aBind->delete("suid = '" . $arrAccessToken['id'] . "' and service = '". $this->params['service']."'") ;
 				$this->createMessage(Message::success,"原有的帐号绑定已经解除。") ;
 			} catch(\Exception $e) {
 				$this->createMessage(Message::error,"系统在解除原有的帐号绑定时遇到错误。") ;
@@ -210,51 +198,56 @@ class AuthoritionObtaining extends UserPanel
 			}
 		}
 		
-		$this->user['token.service'] = $this->params['service'] ;
-		$this->user['token.suid'] = $arrAccessToken['id'] ;
-		$this->user['token.nickname'] = $arrAccessToken['nickname'] ;
-		$this->user['token.username'] = $arrAccessToken['username'] ;
-		$this->user['token.token'] = $arrAccessToken['oauth_token'] ;
-		$this->user['token.token_secret'] = $arrAccessToken['oauth_token_secret'] ;
-		// $this->user->printStruct() ;
-		
 		try{
-			$this->user->save() ;
+			$this->model('oauth:user')->insert(
+	            array(
+	                    'uid' =>$this->model('coresystem:user')->uid ,
+	                    'service' =>$this->params['service'] ,
+	                    'suid' =>$arrAccessToken['id'] ,
+	                    'nickname' =>$arrAccessToken['nickname'] ,
+	                    'username' =>$arrAccessToken['username'] ,
+	                    'token' =>$arrAccessToken['oauth_token'] ,
+	                    'token_secret' =>$arrAccessToken['oauth_token_secret'] ,
+                )
+	        ) ;
+			
 			$this->createMessage(Message::success,"帐号绑定成功") ;			
 		} catch (\Exception $e) {
 			$this->createMessage(Message::error,"帐号绑定失败:%s",$e->getMessage()) ;
 			return ;
 		}
-		
 		$this->location("?c=org.opencomb.oauth.controlPanel.OAuthState") ;
 	}
 	
 	/**
 	 * 绑定一个新帐号
 	 */
-	protected function actionBindAlong()
+	protected function BindAlong()
 	{
 		$arrAccessToken = Session::singleton()->variable($this->params['service'].'.AccessToken') ;
 		
 		// 创建一个新用户
-		$this->user->username = "{$arrAccessToken['id']}@{$this->params['service']}" ;
-		$this->user->password = md5( $this->user->username . time() );
-		$this->user->lastLoginTime = time() ;
-		$this->user->lastLoginIp = $_SERVER['REMOTE_ADDR'] ;
-		$this->user->registerTime = time() ;
-		$this->user->registerIp = $_SERVER['REMOTE_ADDR'] ;
-		$this->user->activeTime = time() ;
-		$this->user->activeIp = $_SERVER['REMOTE_ADDR'] ;
-		$this->user['info.nickname'] = $this->user->username ;
-		$this->user['token.service'] = $this->params['service'] ;
-		$this->user['token.suid'] = $arrAccessToken['id'] ;
-		$this->user['token.nickname'] = $arrAccessToken['nickname'] ;
-		$this->user['token.username'] = $arrAccessToken['username'] ;
-		$this->user['token.token'] = $arrAccessToken['oauth_token'] ;
-		$this->user['token.token_secret'] = $arrAccessToken['oauth_token_secret'] ;
 		
 		try{
-			$this->user->save() ;
+			$this->model('coresystem:user')->insert(
+			        array(
+			                'username' => "{$arrAccessToken['id']}@{$this->params['service']}" ,
+			                'password' => md5( $this->model('coresystem:user')->username . time() ),
+			                'lastLoginTime' => time() ,
+			                'lastLoginIp' => $_SERVER['REMOTE_ADDR'] ,
+			                'registerTime' => time() ,
+			                'registerIp' => $_SERVER['REMOTE_ADDR'] ,
+			                'activeTime' => time() ,
+			                'activeIp' => $_SERVER['REMOTE_ADDR'] ,
+			                'info.nickname' => $this->model('coresystem:user')->username ,
+			                'token.service' => $this->params['service'] ,
+			                'token.suid' => $arrAccessToken['id'] ,
+			                'token.nickname' => $arrAccessToken['nickname'] ,
+			                'token.username' => $arrAccessToken['username'] ,
+			                'token.token' => $arrAccessToken['oauth_token'] ,
+			                'token.token_secret' => $arrAccessToken['oauth_token_secret'] ,
+	                )
+	        ) ;
 		} catch (\Exception $e) {
 			if( $e instanceof ExecuteException and $e->isDuplicate() )
 			{
